refactor : imageFileServiceImpl 교채
1. 기존(imageFileServiceImpl)을 개선(imageFileServiceImpl2)로 변경
2. 기존 구현체 삭제
